(tae): update replay related
